@@ -4,63 +4,66 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.0",
+        "php": "^8.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "api-platform/core": "^2.6",
-        "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2.5",
-        "doctrine/doctrine-migrations-bundle": "^3.2",
-        "doctrine/orm": "^2.11",
-        "gesdinet/jwt-refresh-token-bundle": "^1.0",
-        "lexik/jwt-authentication-bundle": "^2.14",
-        "nelmio/cors-bundle": "^2.2",
-        "phpdocumentor/reflection-docblock": "^5.3",
-        "phpstan/phpdoc-parser": "^1.2",
-        "ramsey/uuid-doctrine": "^1.8",
-        "sensio/framework-extra-bundle": "^6.2",
-        "symfony/apache-pack": "^1.0",
-        "symfony/asset": "5.4.*",
-        "symfony/console": "5.4.*",
-        "symfony/dotenv": "5.4.*",
-        "symfony/expression-language": "5.4.*",
-        "symfony/flex": "^2.1",
-        "symfony/form": "5.4.*",
-        "symfony/framework-bundle": "5.4.*",
-        "symfony/mime": "5.4.*",
-        "symfony/monolog-bundle": "^3.7",
-        "symfony/property-access": "5.4.*",
-        "symfony/property-info": "5.4.*",
-        "symfony/proxy-manager-bridge": "5.4.*",
-        "symfony/runtime": "5.4.*",
-        "symfony/security-bundle": "5.4.*",
-        "symfony/serializer": "5.4.*",
-        "symfony/twig-bundle": "5.4.*",
-        "symfony/validator": "5.4.*",
-        "symfony/webpack-encore-bundle": "^2.0",
+        "api-platform/core": "*",
+        "doctrine/annotations": "*",
+        "doctrine/doctrine-bundle": "*",
+        "doctrine/doctrine-migrations-bundle": "*",
+        "doctrine/orm": "*",
+        "friendsofsymfony/rest-bundle": "*",
+        "gesdinet/jwt-refresh-token-bundle": "*",
+        "lexik/jwt-authentication-bundle": "*",
+        "nelmio/cors-bundle": "*",
+        "phpdocumentor/reflection-docblock": "*",
+        "phpstan/phpdoc-parser": "*",
+        "ramsey/uuid-doctrine": "*",
+        "sensio/framework-extra-bundle": "*",
+        "symfony/apache-pack": "*",
+        "symfony/asset": "*",
+        "symfony/console": "*",
+        "symfony/dotenv": "*",
+        "symfony/expression-language": "*",
+        "symfony/flex": "*",
+        "symfony/form": "*",
+        "symfony/framework-bundle": "*",
+        "symfony/mime": "*",
+        "symfony/monolog-bundle": "*",
+        "symfony/property-access": "*",
+        "symfony/property-info": "*",
+        "symfony/proxy-manager-bridge": "*",
+        "symfony/runtime": "*",
+        "symfony/security-bundle": "*",
+        "symfony/serializer": "*",
+        "symfony/twig-bundle": "*",
+        "symfony/validator": "*",
+        "symfony/webpack-encore-bundle": "*",
         "symfony/website-skeleton": "*",
-        "symfony/yaml": "5.4.*",
-        "thecodingmachine/safe": "^2.1",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0"
+        "symfony/yaml": "*",
+        "thecodingmachine/safe": "*",
+        "twig/extra-bundle": "*",
+        "twig/twig": "*"
     },
     "require-dev": {
-        "doctrine/coding-standard": "^9.0",
+        "doctrine/coding-standard": "*",
         "doctrine/doctrine-fixtures-bundle": "^3.4",
-        "friendsoftwig/twigcs": "^5.1",
-        "phpstan/phpstan": "^1.4",
-        "phpunit/phpunit": "^9.5",
-        "squizlabs/php_codesniffer": "^3.6",
-        "symfony/browser-kit": "5.4.*",
-        "symfony/css-selector": "5.4.*",
-        "symfony/debug-bundle": "5.4.*",
-        "symfony/maker-bundle": "^1.36",
-        "symfony/phpunit-bridge": "^6.0",
-        "symfony/process": "5.4.*",
-        "symfony/stopwatch": "5.4.*",
-        "symfony/web-profiler-bundle": "5.4.*",
-        "thecodingmachine/phpstan-safe-rule": "^1.2",
-        "thecodingmachine/phpstan-strict-rules": "^1.0"
+        "friendsoftwig/twigcs": "*",
+        "phpstan/phpstan": "*",
+        "phpunit/phpunit": "*",
+        "rector/rector": "*",
+        "squizlabs/php_codesniffer": "*",
+        "symfony/browser-kit": "6.3.*",
+        "symfony/css-selector": "*",
+        "symfony/debug-bundle": "*",
+        "symfony/http-client": "6.3.*",
+        "symfony/maker-bundle": "*",
+        "symfony/phpunit-bridge": "*",
+        "symfony/process": "*",
+        "symfony/stopwatch": "*",
+        "symfony/web-profiler-bundle": "*",
+        "thecodingmachine/phpstan-safe-rule": "*",
+        "thecodingmachine/phpstan-strict-rules": "*"
     },
     "config": {
         "allow-plugins": {
