@@ -5,10 +5,8 @@ namespace App\Entity;
 use Doctrine\ORM\Mapping as ORM;
 use Gesdinet\JWTRefreshTokenBundle\Entity\RefreshToken as BaseRefreshToken;
 
-/**
- * @ORM\Entity
- * @ORM\Table("refresh_tokens")
- */
+#[ORM\Table('refresh_tokens')]
+#[ORM\Entity]
 class RefreshToken extends BaseRefreshToken
 {
 }
