@@ -2,47 +2,46 @@
 
 namespace App\Entity;
 
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\ApiResource;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
 
-/*#[ApiResource(formats: ['xml', 'json', 'csv' => ['text/csv']])]*/
-#[ApiResource(formats: ['json'])]
+/*#[ApiResource(formats: ['xml', 'json', 'ld+json', 'csv' => ['text/csv']])]*/
+#[ApiResource(formats: ['json', 'jsonld'])]
 /**
  * Contact
- *
- * @ORM\Table(name="contact")
- * @ORM\Entity
  */
+#[Orm\Table("contact")]
+#[Orm\Entity()]
 class Contacts
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer", nullable=false)
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
      */
+    #[ORM\Column(name: 'id', type: 'integer', nullable: false)]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
     private $id;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="first_name", type="string", length=255, nullable=false)
      */
+    #[ORM\Column(name: 'first_name', type: 'string', length: 255, nullable: false)]
+    #[Assert\NotBlank()]
     private $firstName;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="last_name", type="string", length=255, nullable=false)
      */
+    #[ORM\Column(name: 'last_name', type: 'string', length: 255, nullable: false)]
+    #[Assert\NotBlank()]
     private $lastName;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="email_address", type="string", length=255, nullable=false)
      */
+    #[ORM\Column(name: 'email_address', type: 'string', length: 255, nullable: false)]
+    #[Assert\NotBlank()]
     private $emailAddress;
 
     public function getId(): ?int
