@@ -2,14 +2,18 @@
 
 namespace App\Entity;
 
-use App\Controller\DevicePreviewPictureController;
 use App\Controller\DeviceImageController;
 use App\Controller\DeviceImageUploadController;
 use App\Controller\DeviceImageDeleteController;
 use App\Controller\UploadImageDeleteController;
 use App\Entity\DeviceXDeviceOption;
-use ApiPlatform\Core\Annotation\ApiProperty;
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\ApiResource;
+use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\Get;
+use ApiPlatform\Metadata\GetCollection;
+use ApiPlatform\Metadata\Patch;
+use ApiPlatform\Metadata\Post;
+use ApiPlatform\Metadata\Put;
 use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
@@ -19,196 +23,154 @@ use Doctrine\ORM\Mapping as ORM;
 
 /**
  * Devices
- *
- * @ORM\Table(
- *    name="devices",
- *    uniqueConstraints={
- *      @ORM\UniqueConstraint(name="unique_device_name", columns={"name"}),
- *      @ORM\UniqueConstraint(name="unique_device_seo_link", columns={"seo_link"})
- *    },
- *    indexes={
- *      @ORM\Index(name="device_creator", columns={"creator"}),
- *      @ORM\Index(name="device_id", columns={"id"}),
- *      @ORM\Index(name="device_updater", columns={"updater"})
- *    }
- * )
- * @ORM\Entity
- * @ORM\HasLifecycleCallbacks()
- * @ApiResource(
- *   attributes={"validation_groups"={"write"}},
- *   normalizationContext={"groups"={"read"}},
- *   denormalizationContext={"groups"={"write"}},
- *   collectionOperations={
- *     "get"={"method"="GET"},
- *     "post"={"method"="POST", "validation_groups"={"write"}},
- *     "get_images_for_device" = {
- *        "method" = "GET",
- *        "path" = "/devices/{id}/images",
- *        "controller" = DeviceImageController::class,
- *        "read" = false,
- *        "openapi_context" = {
- *          "parameters" = {
- *           {
- *             "name" = "id",
- *             "in" = "path",
- *             "description" = "id of device",
- *             "type" = "integer",
- *             "required" = false,
- *             "example"= 1,
- *           },
- *         },
- *       },
- *     },
- *     "delete_upload_image" = {
- *       "method" = "DELETE",
- *       "path" = "/uploads/image/{fileName}",
- *       "controller" = UploadImageDeleteController::class
- *     },
- *     "delete_device_image" = {
- *       "method" = "DELETE",
- *       "path" = "/devices/{id}/image/{fileName}",
- *       "controller" = DeviceImageDeleteController::class
- *     },
- *     "post_upload_images" = {
- *       "method" = "POST",
- *       "path" = "/devices/images",
- *       "controller" = DeviceImageUploadController::class,
- *       "deserialize" = false,
- *       "openapi_context" = {
- *         "requestBody" = {
- *           "description" = "File upload to an existing resource (device)",
- *           "required" = true,
- *           "content" = {
- *             "multipart/form-data" = {
- *               "schema" = {
- *                 "type" = "object",
- *                 "properties" = {
- *                   "file" = {
- *                     "type" = "string",
- *                     "format" = "binary",
- *                     "description" = "Upload images for device",
- *                   },
- *                 },
- *               },
- *             },
- *           },
- *         },
- *       },
- *     }
- *   },
- *   itemOperations={
- *     "get",
- *     "delete",
- *     "put"={"method"="PUT", "validation_groups"={"write"}},
- *     "get_by_slug" = {
- *       "method" = "GET",
- *       "path" = "/superhero/{slug}",
- *       "controller" = SuperheroBySlug::class,
- *       "read"=false,
- *       "openapi_context" = {
- *         "parameters" = {
- *           {
- *             "name" = "slug",
- *             "in" = "path",
- *             "description" = "The slug of your superhero",
- *             "type" = "string",
- *             "required" = true,
- *             "example"= "superman",
- *           },
- *         },
- *       },
- *     },
- *   }
- * )
  */
+#[ApiResource(
+  validationContext: ['groups' => ['device:write']],
+  normalizationContext: ['groups' => ['device:read']],
+  denormalizationContext: ['groups' => ['device:write']],
+  operations: [
+    new Get(),
+    new Get(
+      uriTemplate: "/devices/{id}/images",
+      requirements: ['id' => '\d+'],
+      controller: DeviceImageController::class,
+      read: false,
+      openapiContext: [
+        'parameters' => [
+          [
+            "name" => "id",
+            "in" => "path",
+            "description" => "id of device",
+            "type" => "integer",
+            "required" => true,
+            "example"=> 1,
+          ],
+        ],
+      ],
+    ),
+    new GetCollection(),
+    new Post(),
+    new Post(
+      uriTemplate: "/devices/images",
+      controller: DeviceImageUploadController::class,
+      deserialize: false,
+      openapiContext: [
+        "requestBody" => [
+          "description" => "File upload to an existing resource (device)",
+          "required" => true,
+          "content" => [
+            "multipart/form-data" => [
+              "schema" => [
+                "type" => "object",
+                "properties" => [
+                  "file" => [
+                    "type" => "string",
+                    "format" => "binary",
+                    "description" => "image path name for device"
+                  ]
+                ]
+              ]
+            ]
+          ]
+        ]
+      ]
+    ),
+    new Patch(),
+    new Put(),
+    new Delete(),
+    new Delete(
+      uriTemplate: "/uploads/image/{fileName}",
+      controller: UploadImageDeleteController::class
+    ),
+    new Delete(
+      uriTemplate: "/devices/{id}/image/{fileName}",
+      controller: DeviceImageDeleteController::class
+    )
+  ]
+)]
+#[ORM\Table(name: 'devices')]
+#[ORM\Index(name: 'device_id', columns: ['id'])]
+#[ORM\Index(name: 'device_creator', columns: ['creator'])]
+#[ORM\Index(name: 'device_updater', columns: ['updater'])]
+#[ORM\UniqueConstraint(name: 'unique_device_name', columns: ['name'])]
+#[ORM\UniqueConstraint(name: 'unique_device_seo_link', columns: ['seo_link'])]
+#[ORM\Entity]
+#[ORM\HasLifecycleCallbacks]
 class Devices
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer", nullable=false, options={"unsigned"=true})
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'id', type: 'integer', nullable: false, options: ['unsigned' => true])]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
+    #[Groups(['device:read', 'device:write'])]
     private $id;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="name", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
-     * @Assert\NotBlank(
-     *    message="This value should not be blank",
-     *    groups={"write"}
-     * )
      */
+    #[ORM\Column(name: 'name', type: 'string', length: 250, nullable: false)]
+    #[Assert\NotBlank(groups: [ 'device:write' ])]
+    #[Groups(['device:read', 'device:write'])]
     private $name;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="seo_link", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'seo_link', type: 'string', length: 250, nullable: false)]
+    #[Groups(['device:read', 'device:write'])]
     private $seoLink;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="preview_picture_path", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'preview_picture_path', type: 'string', length: 250, nullable: false)]
+    #[Groups(['device:read', 'device:write'])]
     private $previewPicturePath;
 
     /**
      * @var Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumn(name="creator", referencedColumnName="id")
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: 'Users')]
+    #[ORM\JoinColumn(name: 'creator', referencedColumnName: 'id')]
+    #[Groups(['device:read'])]
     private $creator;
 
     /**
      * @var \DateTime
-     *
-     * @ORM\Column(name="created", type="datetime", nullable=false, options={"default"="CURRENT_TIMESTAMP"})
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'created', type: 'datetime', nullable: false, options: ['default' => 'CURRENT_TIMESTAMP'])]
+    #[Groups(['device:read'])]
     private $created = 'CURRENT_TIMESTAMP';
 
     /**
      * @var Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumn(name="updater", referencedColumnName="id")
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: 'Users')]
+    #[ORM\JoinColumn(name: 'updater', referencedColumnName: 'id')]
+    #[Groups(['device:read'])]
     private $updater;
 
     /**
      * @var \DateTime|null
-     *
-     * @ORM\Column(name="updated", type="datetime", nullable=true)
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'updated', type: 'datetime', nullable: true)]
+    #[Groups(['device:read'])]
     private $updated;
 
     /**
      * @var Collection|DeviceOption[]
-     *
-     * @ORM\OneToMany(targetEntity=DeviceXDeviceOption::class, mappedBy="device", cascade={"ALL"}, orphanRemoval=true)
-     * @Groups({"read", "write"})
      */
+    #[ORM\OneToMany(targetEntity: DeviceXDeviceOption::class, mappedBy: 'device', cascade: ['ALL'], orphanRemoval: true)]
+    #[Groups(['device:read', 'device:write'])]
     private $deviceXDeviceOptions;
 
     /**
      * @var Collection|DeviceOption[]
-     *
-     * @ORM\OneToMany(targetEntity=DeviceXDeviceGroup::class, mappedBy="device", cascade={"ALL"}, orphanRemoval=true)
-     * @Groups({"read", "write"})
      */
+    #[ORM\OneToMany(targetEntity: DeviceXDeviceGroup::class, mappedBy: 'device', cascade: ['ALL'], orphanRemoval: true)]
+    #[Groups(['device:read', 'device:write'])]
     private $deviceXDeviceGroups;
 
     public function __construct()
