@@ -6,8 +6,14 @@ use App\Controller\ExerciseImageController;
 use App\Controller\ExerciseImageDeleteController;
 use App\Controller\ExerciseImageUploadController;
 use App\Controller\UploadImageDeleteController;
-use ApiPlatform\Core\Annotation\ApiProperty;
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\ApiProperty;
+use ApiPlatform\Metadata\ApiResource;
+use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\Get;
+use ApiPlatform\Metadata\GetCollection;
+use ApiPlatform\Metadata\Patch;
+use ApiPlatform\Metadata\Post;
+use ApiPlatform\Metadata\Put;
 use DateTime;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
@@ -16,158 +22,121 @@ use Doctrine\Common\Collections\Collection;
 
 /**
  * Exercises
- *
- * @ORM\Table(
- *    name="exercises",
- *    uniqueConstraints={
- *        @ORM\UniqueConstraint(name="UN_exercise_name", columns={"name"}),
- *        @ORM\UniqueConstraint(name="UN_exercise_seo_link", columns={"seo_link"})
- *    },
- *    indexes={
- *        @ORM\Index(name="IDX_exercise_creator", columns={"creator"}),
- *        @ORM\Index(name="IDX_exercise_id", columns={"id"}),
- *        @ORM\Index(name="IDX_exercise_updater", columns={"updater"})
- *    }
- * )
- * @ORM\Entity
- * @ORM\HasLifecycleCallbacks()
- * @ApiResource(
- *   normalizationContext={"groups" = {"read"}},
- *   denormalizationContext={"groups" = {"write"}},
- *   collectionOperations={
- *     "get",
- *     "post",
- *     "get_images_for_exercise" = {
- *        "method" = "GET",
- *        "path" = "/exercises/{id}/images",
- *        "controller" = ExerciseImageController::class,
- *        "read" = false,
- *        "openapi_context" = {
- *          "parameters" = {
- *           {
- *             "name" = "id",
- *             "in" = "path",
- *             "description" = "id of exercise",
- *             "type" = "integer",
- *             "required" = false,
- *             "example"= 1,
- *           },
- *         },
- *       },
- *     },
- *     "delete_upload_image" = {
- *       "method" = "DELETE",
- *       "path" = "/uploads/image/{fileName}",
- *       "controller" = UploadImageDeleteController::class
- *     },
- *     "delete_exercise_image" = {
- *       "method" = "DELETE",
- *       "path" = "/exercises/{id}/image/{fileName}",
- *       "controller" = ExerciseImageDeleteController::class
- *     },
- *     "post_upload_images" = {
- *       "method" = "POST",
- *       "path" = "/exercises/images",
- *       "controller" = ExerciseImageUploadController::class,
- *       "deserialize" = false,
- *       "openapi_context" = {
- *         "requestBody" = {
- *           "description" = "File upload to an existing resource (exercise)",
- *           "required" = true,
- *           "content" = {
- *             "multipart/form-data" = {
- *               "schema" = {
- *                 "type" = "object",
- *                 "properties" = {
- *                   "file" = {
- *                     "type" = "string",
- *                     "format" = "binary",
- *                     "description" = "Upload images for exercise",
- *                   },
- *                 },
- *               },
- *             },
- *           },
- *         },
- *       },
- *     }
- *   },
- *   itemOperations={
- *     "get",
- *     "patch",
- *     "delete",
- *     "put",
- *     "get_by_slug" = {
- *       "method" = "GET",
- *       "path" = "/superhero/{slug}",
- *       "controller" = SuperheroBySlug::class,
- *       "read"=false,
- *       "openapi_context" = {
- *         "parameters" = {
- *           {
- *             "name" = "slug",
- *             "in" = "path",
- *             "description" = "The slug of your superhero",
- *             "type" = "string",
- *             "required" = true,
- *             "example"= "superman",
- *           },
- *         },
- *       },
- *     },
- *   }
- * )
  */
+#[ApiResource(
+  normalizationContext: ['groups' => ['read']],
+  denormalizationContext: ['groups' => ['write']],
+  operations: [
+    new Get(),
+    new Get(
+      uriTemplate: "/exercises/{id}/images",
+      requirements: ['id' => '\d+'],
+      controller: ExerciseImageController::class,
+      read: false,
+      openapiContext: [
+        'parameters' => [
+          [
+            "name" => "id",
+            "in" => "path",
+            "description" => "id of exercise",
+            "type" => "integer",
+            "required" => true,
+            "example"=> 1,
+          ],
+        ],
+      ],
+    ),
+    new GetCollection(),
+    new Post(),
+    new Post(
+      uriTemplate: "/exercises/images",
+      controller: ExerciseImageUploadController::class,
+      deserialize: false,
+      openapiContext: [
+        "requestBody" => [
+          "description" => "File upload to an existing resource (exercise)",
+          "required" => true,
+          "content" => [
+            "multipart/form-data" => [
+              "schema" => [
+                "type" => "object",
+                "properties" => [
+                  "file" => [
+                    "type" => "string",
+                    "format" => "binary",
+                    "description" => "image path name for exercise"
+                  ]
+                ]
+              ]
+            ]
+          ]
+        ]
+      ]
+    ),
+    new Patch(),
+    new Put(),
+    new Delete(),
+    new Delete(
+      uriTemplate: "/uploads/image/{fileName}",
+      controller: UploadImageDeleteController::class
+    ),
+    new Delete(
+      uriTemplate: "/exercises/{id}/image/{fileName}",
+      controller: ExerciseImageDeleteController::class
+    )
+  ]
+)]
+#[ORM\Table(name: 'exercises')]
+#[ORM\Index(name: 'IDX_exercise_creator', columns: ['creator'])]
+#[ORM\Index(name: 'IDX_exercise_id', columns: ['id'])]
+#[ORM\Index(name: 'IDX_exercise_updater', columns: ['updater'])]
+#[ORM\UniqueConstraint(name: 'UN_exercise_name', columns: ['name'])]
+#[ORM\UniqueConstraint(name: 'UN_exercise_seo_link', columns: ['seo_link'])]
+#[ORM\Entity]
+#[ORM\HasLifecycleCallbacks]
 class Exercises
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer", nullable=false, options={"unsigned"=true})
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'id', type: 'integer', nullable: false, options: ['unsigned' => true])]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
+    #[Groups(['read', 'write'])]
     private $id;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="name", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'name', type: 'string', length: 250, nullable: false)]
+    #[Groups(['read', 'write'])]
     private $name;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="seo_link", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'seo_link', type: 'string', length: 250, nullable: false)]
+    #[Groups(['read', 'write'])]
     private $seoLink;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="description", type="text", length=65535, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'description', type: 'text', length: 65535, nullable: false)]
+    #[Groups(['read', 'write'])]
     private $description;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="special_features", type="text", length=65535, nullable=false, options={"comment"="Besonderheiten"})
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'special_features', type: 'text', length: 65535, nullable: false, options: ['comment' => 'Besonderheiten'])]
+    #[Groups(['read', 'write'])]
     private $specialFeatures;
 
     /**
      * @var string
      *
-     * @ORM\Column(name="preview_picture_path", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
-     * @ApiProperty(
+     * #[ApiProperty(
      *   iri="http://schema.org/image",
      *   attributes={
      *     "openapi_context"={
@@ -175,73 +144,68 @@ class Exercises
      *     }
      *   }
      * )
+     * ]
      */
+    #[ORM\Column(name: 'preview_picture_path', type: 'string', length: 250, nullable: false)]
+    #[Groups(['read', 'write'])]
     private $previewPicturePath;
 
     /**
      * @var \DateTime
-     *
-     * @ORM\Column(name="created", type="datetime", nullable=false, options={"default"="CURRENT_TIMESTAMP"})
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'created', type: 'datetime', nullable: false, options: ['default' => 'CURRENT_TIMESTAMP'])]
+    #[Groups(['read'])]
     private $created = 'CURRENT_TIMESTAMP';
 
     /**
      * @var \DateTime|null
-     *
-     * @ORM\Column(name="updated", type="datetime", nullable=true)
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'updated', type: 'datetime', nullable: true)]
+    #[Groups(['read'])]
     private $updated;
 
     /**
      * @var Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumn(name="creator", referencedColumnName="id")
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: 'Users')]
+    #[ORM\JoinColumn(name: 'creator', referencedColumnName: 'id')]
+    #[Groups(['read'])]
     private $creator;
 
     /**
      * @var Users
-     *
-     * @ORM\ManyToOne(targetEntity=Users::class)
-     * @ORM\JoinColumn(name="updater", referencedColumnName="id", nullable=true)
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: Users::class)]
+    #[ORM\JoinColumn(name: 'updater', referencedColumnName: 'id', nullable: true)]
+    #[Groups(['read'])]
     private $updater;
 
     /**
      * @var Collection|ExerciseOption[]
-     *
-     * @ORM\OneToMany(targetEntity=ExerciseXExerciseOption::class, mappedBy="exercise", cascade={"ALL"}, orphanRemoval=true)
-     * @Groups({"read", "write"})
      */
+    #[ORM\OneToMany(targetEntity: ExerciseXExerciseOption::class, mappedBy: 'exercise', cascade: ['ALL'], orphanRemoval: true)]
+    #[Groups(['read', 'write'])]
     private Collection $exerciseXExerciseOptions;
 
     /**
      * @var Collection|DeviceOption[]
-     *
-     * @ORM\OneToMany(targetEntity=ExerciseXDeviceOption::class, mappedBy="exercise", cascade={"ALL"}, orphanRemoval=true)
-     * @Groups({"read", "write"})
      */
+    #[ORM\OneToMany(targetEntity: ExerciseXDeviceOption::class, mappedBy: 'exercise', cascade: ['ALL'], orphanRemoval: true)]
+    #[Groups(['read', 'write'])]
     private Collection $exerciseXDeviceOptions;
 
     /**
      * @var ArrayCollection|Device[]
-     *
-     * @ORM\OneToMany(targetEntity=ExerciseXDevice::class, mappedBy="exercise", cascade={"ALL"}, orphanRemoval=true)
-     * @Groups({"read", "write"})
      */
+    #[ORM\OneToMany(targetEntity: ExerciseXDevice::class, mappedBy: 'exercise', cascade: ['ALL'], orphanRemoval: true)]
+    #[Groups(['read', 'write'])]
     private Collection $exerciseXDevices;
 
     /**
      * @var ArrayCollection|ExerciseType[]
-     *
-     * @ORM\OneToMany(targetEntity=ExerciseXExerciseType::class, mappedBy="exercise", cascade={"ALL"}, orphanRemoval=true)
-     * @Groups({"read", "write"})
      */
+    #[ORM\OneToMany(targetEntity: ExerciseXExerciseType::class, mappedBy: 'exercise', cascade: ['ALL'], orphanRemoval: true)]
+    #[Groups(['read', 'write'])]
     private Collection $exerciseXExerciseType;
 
     public function __construct()
