@@ -4,8 +4,14 @@ namespace App\Entity;
 
 use App\Entity\MuscleGroups;
 use App\Entity\MuscleXMuscleGroup;
-use ApiPlatform\Core\Annotation\ApiProperty;
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\ApiProperty;
+use ApiPlatform\Metadata\ApiResource;
+use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\Get;
+use ApiPlatform\Metadata\GetCollection;
+use ApiPlatform\Metadata\Patch;
+use ApiPlatform\Metadata\Post;
+use ApiPlatform\Metadata\Put;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Serializer\Annotation\Groups;
@@ -13,103 +19,87 @@ use Doctrine\Common\Collections\Collection;
 
 /**
  * Muscles
- *
- * @ORM\Table(
- *    name="muscles",
- *    uniqueConstraints={
- *      @ORM\UniqueConstraint(name="muscle_name", columns={"name"}),
- *      @ORM\UniqueConstraint(name="muscle_seo_link", columns={"seo_link"})
- *    },
- *    indexes={
- *      @ORM\Index(name="muscle_creator", columns={"creator"}),
- *      @ORM\Index(name="muscle_id", columns={"id"}),
- *      @ORM\Index(name="muscle_updater", columns={"updater"})
- *    }
- * )
- * @ORM\Entity
- * @ORM\HasLifecycleCallbacks()
- * @ApiResource(
- *   normalizationContext={"groups" = {"read"}},
- *   denormalizationContext={"groups" = {"write"}},
- *   itemOperations={
- *     "get"={"method"="GET"},
- *     "put"={"method"="PUT"},
- *     "delete"={"method"="DELETE"}
- *   },
- *   collectionOperations={
- *     "get"={"method"="GET"},
- *     "post"={"method"="POST"}
- *   }
- * )
  */
+#[ApiResource(
+  normalizationContext: ['groups' => ['read']],
+  denormalizationContext: ['groups' => ['write']],
+  operations: [
+    new Get(),
+    new GetCollection(),
+    new Post(),
+    new Patch(),
+    new Put(),
+    new Delete()
+  ]
+)]
+#[ORM\Table(name: 'muscles')]
+#[ORM\Index(name: 'muscle_creator', columns: ['creator'])]
+#[ORM\Index(name: 'muscle_id', columns: ['id'])]
+#[ORM\Index(name: 'muscle_updater', columns: ['updater'])]
+#[ORM\UniqueConstraint(name: 'muscle_name', columns: ['name'])]
+#[ORM\UniqueConstraint(name: 'muscle_seo_link', columns: ['seo_link'])]
+#[ORM\Entity]
+#[ORM\HasLifecycleCallbacks]
 class Muscles
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer", nullable=false, options={"unsigned"=true})
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'id', type: 'integer', nullable: false, options: ['unsigned' => true])]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
+    #[Groups(['read', 'write'])]
     private $id;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="name", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'name', type: 'string', length: 250, nullable: false)]
+    #[Groups(['read', 'write'])]
     private $name;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="seo_link", type="string", length=250, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'seo_link', type: 'string', length: 250, nullable: false)]
+    #[Groups(['read', 'write'])]
     private $seoLink;
 
     /**
      * @var \DateTime
-     *
-     * @ORM\Column(name="created", type="datetime", nullable=false, options={"default"="CURRENT_TIMESTAMP"})
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'created', type: 'datetime', nullable: false, options: ['default' => 'CURRENT_TIMESTAMP'])]
+    #[Groups(['read'])]
     private $created = 'CURRENT_TIMESTAMP';
 
     /**
      * @var \DateTime|null
-     *
-     * @ORM\Column(name="updated", type="datetime", nullable=true)
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'updated', type: 'datetime', nullable: true)]
+    #[Groups(['read'])]
     private $updated;
 
     /**
      * @var Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumn(name="creator", referencedColumnName="id")
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: 'Users')]
+    #[ORM\JoinColumn(name: 'creator', referencedColumnName: 'id')]
+    #[Groups(['read'])]
     private $creator;
 
     /**
      * @var Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumn(name="updater", referencedColumnName="id")
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: 'Users')]
+    #[ORM\JoinColumn(name: 'updater', referencedColumnName: 'id')]
+    #[Groups(['read'])]
     private $updater;
 
     /**
      * @var Collection|MuscleGroups[]
-     *
-     * @ORM\OneToMany(targetEntity=MuscleXMuscleGroup::class, mappedBy="muscle", cascade={"ALL"}, orphanRemoval=true)
-     * @Groups({"read", "write"})
      */
+    #[ORM\OneToMany(targetEntity: MuscleXMuscleGroup::class, mappedBy: 'muscle', cascade: ['ALL'], orphanRemoval: true)]
+    #[Groups(['read', 'write'])]
     private $muscleXMuscleGroups;
 
     public function __construct()
