@@ -2,101 +2,83 @@
 
 namespace App\Entity;
 
-use ApiPlatform\Core\Annotation\ApiResource;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Serializer\Annotation\Groups;
 
 /**
  * TrainingPlanXDeviceOption
- *
- * @ORM\Table(name="training_plan_x_device_option", uniqueConstraints={@ORM\UniqueConstraint(name="unique_training_plan_x_device_option_id", columns={"id"})}, indexes={@ORM\Index(name="training_plan_x_device_option_create_user_fk", columns={"creator"}), @ORM\Index(name="training_plan_x_device_option_device_option_fk", columns={"device_option"}), @ORM\Index(name="training_plan_x_device_option_id", columns={"id"}), @ORM\Index(name="training_plan_x_device_option_training_plan_exercise_fk", columns={"training_plan_x_exercise"}), @ORM\Index(name="training_plan_x_device_option_update_user_fk", columns={"updater"})})
- * @ORM\Entity
- * @ORM\HasLifecycleCallbacks()
- * @ApiResource(
- *   itemOperations={
- *     "get",
- *     "patch",
- *     "delete",
- *     "put"
- *   },
- *   normalizationContext={"groups" = {"read"}},
- *   denormalizationContext={"groups" = {"write"}},
- *   collectionOperations={
- *     "get",
- *     "post"
- *   }
- * )
  */
+#[ORM\Table(name: 'training_plan_x_device_option')]
+#[ORM\Index(name: 'training_plan_x_device_option_create_user_fk', columns: ['creator'])]
+#[ORM\Index(name: 'training_plan_x_device_option_device_option_fk', columns: ['device_option'])]
+#[ORM\Index(name: 'training_plan_x_device_option_id', columns: ['id'])]
+#[ORM\Index(name: 'training_plan_x_device_option_training_plan_exercise_fk', columns: ['training_plan_x_exercise'])]
+#[ORM\Index(name: 'training_plan_x_device_option_update_user_fk', columns: ['updater'])]
+#[ORM\UniqueConstraint(name: 'unique_training_plan_x_device_option_id', columns: ['id'])]
+#[ORM\Entity]
+#[ORM\HasLifecycleCallbacks]
 class TrainingPlanXDeviceOption
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer", nullable=false, options={"unsigned"=true})
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'id', type: 'integer', nullable: false, options: ['unsigned' => true])]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
+    #[Groups(['read', 'write'])]
     private $id;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="option_value", type="string", length=255, nullable=false)
-     * @Groups({"read", "write"})
      */
+    #[ORM\Column(name: 'option_value', type: 'string', length: 255, nullable: false)]
+    #[Groups(['read', 'write'])]
     private $optionValue;
 
     /**
      * @var \DeviceOptions
-     *
-     * @ORM\ManyToOne(targetEntity="DeviceOptions")
-     * @ORM\JoinColumn(name="device_option", referencedColumnName="id")
-     * @Groups({"read", "write"})
      */
+    #[ORM\ManyToOne(targetEntity: 'DeviceOptions')]
+    #[ORM\JoinColumn(name: 'device_option', referencedColumnName: 'id')]
+    #[Groups(['read', 'write'])]
     private $deviceOption;
 
     /**
      * @var \TrainingPlanXExercise
-     *
-     * @ORM\ManyToOne(targetEntity="TrainingPlanXExercise")
-     * @ORM\JoinColumn(name="training_plan_x_exercise", referencedColumnName="id")
-     * @Groups({"read", "write"})
      */
+    #[ORM\ManyToOne(targetEntity: 'TrainingPlanXExercise')]
+    #[ORM\JoinColumn(name: 'training_plan_x_exercise', referencedColumnName: 'id')]
+    #[Groups(['read', 'write'])]
     private $trainingPlanXExercise;
 
     /**
      * @var \Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumn(name="creator", referencedColumnName="id")
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: 'Users')]
+    #[ORM\JoinColumn(name: 'creator', referencedColumnName: 'id')]
+    #[Groups(['read'])]
     private $creator;
 
     /**
      * @var \DateTime
-     *
-     * @ORM\Column(name="created", type="datetime", nullable=false, options={"default"="CURRENT_TIMESTAMP"})
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'created', type: 'datetime', nullable: false, options: ['default' => 'CURRENT_TIMESTAMP'])]
+    #[Groups(['read'])]
     private $created = 'CURRENT_TIMESTAMP';
 
     /**
      * @var \Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumn(name="updater", referencedColumnName="id")
-     * @Groups({"read"})
      */
+    #[ORM\ManyToOne(targetEntity: 'Users')]
+    #[ORM\JoinColumn(name: 'updater', referencedColumnName: 'id')]
+    #[Groups(['read'])]
     private $updater;
 
     /**
      * @var \DateTime|null
-     *
-     * @ORM\Column(name="updated", type="datetime", nullable=true)
-     * @Groups({"read"})
      */
+    #[ORM\Column(name: 'updated', type: 'datetime', nullable: true)]
+    #[Groups(['read'])]
     private $updated;
 
     /**
