@@ -6,67 +6,58 @@ use Doctrine\ORM\Mapping as ORM;
 
 /**
  * Widgets
- *
- * @ORM\Table(name="widgets", indexes={@ORM\Index(name="widget_create_user_fk", columns={"creator"}), @ORM\Index(name="widget_id", columns={"id"}), @ORM\Index(name="widget_update_user_fk", columns={"updater"})})
- * @ORM\Entity
  */
+#[ORM\Table(name: 'widgets')]
+#[ORM\Index(name: 'widget_create_user_fk', columns: ['creator'])]
+#[ORM\Index(name: 'widget_id', columns: ['id'])]
+#[ORM\Index(name: 'widget_update_user_fk', columns: ['updater'])]
+#[ORM\Entity]
 class Widgets
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer", nullable=false, options={"unsigned"=true})
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
      */
+    #[ORM\Column(name: 'id', type: 'integer', nullable: false, options: ['unsigned' => true])]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
     private $id;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="name", type="string", length=50, nullable=false)
      */
+    #[ORM\Column(name: 'name', type: 'string', length: 50, nullable: false)]
     private $name;
 
     /**
      * @var bool
-     *
-     * @ORM\Column(name="editable", type="boolean", nullable=false)
      */
+    #[ORM\Column(name: 'editable', type: 'boolean', nullable: false)]
     private $editable = '0';
 
     /**
      * @var \DateTime
-     *
-     * @ORM\Column(name="created", type="datetime", nullable=false, options={"default"="CURRENT_TIMESTAMP"})
      */
+    #[ORM\Column(name: 'created', type: 'datetime', nullable: false, options: ['default' => 'CURRENT_TIMESTAMP'])]
     private $created = 'CURRENT_TIMESTAMP';
 
     /**
      * @var \DateTime|null
-     *
-     * @ORM\Column(name="updated", type="datetime", nullable=true)
      */
+    #[ORM\Column(name: 'updated', type: 'datetime', nullable: true)]
     private $updated;
 
     /**
      * @var \Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="creator", referencedColumnName="id")
-     * })
      */
+    #[ORM\JoinColumn(name: 'creator', referencedColumnName: 'id')]
+    #[ORM\ManyToOne(targetEntity: 'Users')]
     private $creator;
 
     /**
      * @var \Users
-     *
-     * @ORM\ManyToOne(targetEntity="Users")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="updater", referencedColumnName="id")
-     * })
      */
+    #[ORM\JoinColumn(name: 'updater', referencedColumnName: 'id')]
+    #[ORM\ManyToOne(targetEntity: 'Users')]
     private $updater;
 
 
